+++ src/eca/llm_providers/openai.clj
@@ -72,14 +72,14 @@
                      (llm-util/log-response logger-tag rid "response-error" body)
                      {:error error-data}))]
     (llm-util/log-request logger-tag rid url body headers)
-    (let [{:keys [status body]} (http/post
-                                 url
-                                 {:headers headers
-                                  :body (json/generate-string body)
-                                  :throw-exceptions? false
-                                  :http-client (client/merge-with-global-http-client http-client)
-                                  :as (if on-stream :stream :json)})]
-      (try
+    (try
+      (let [{:keys [status body]} (http/post
+                                   url
+                                   {:headers headers
+                                    :body (json/generate-string body)
+                                    :throw-exceptions? false
+                                    :http-client (client/merge-with-global-http-client http-client)
+                                    :as (if on-stream :stream :json)})]
         (if (not= 200 status)
           (let [body-str (if on-stream (slurp body) body)]
             (logger/warn logger-tag "Unexpected response status: %s body: %s" status body-str)
@@ -93,9 +93,10 @@
                 (on-stream event data)))
             (do
               (llm-util/log-response logger-tag rid "response" body)
-              (response-body->result body))))
-        (catch Exception e
-          (on-error {:exception e}))))))
+              (response-body->result body)))))
+      (catch Exception e
+        (on-error {:exception e
+                   :message (format "Connection error: %s" (or (ex-message e) (.getName (class e))))})))))
 
 (defn ^:private normalize-messages [messages supports-image?]
   (keep (fn [{:keys [role content] :as msg}]